requirements.txt
firecrawl-py>=1.2.0
groq>=0.13.0
openai>=1.55.0
sarvamai>=0.1.6
python-dotenv>=1.0.1
requests>=2.32.3
fire>=0.6.0
python-docx>=1.1.2
reportlab>=4.2.5
streamlit>=1.37.0
chainlit>=2.0.0
plotly>=5.24.1
